src/requirements.txt
githubpy==1.1.0
agithub==2.1
